Added model for classic map layout
I keep forgetting to push this, but this is basically what I was
envisioning for the classic map for our game. I have not put it in Unity
yet to see if it works as well as I think it does. I fear of collision
issues, but hopefully it won't be too hard to fix if it becomes a
problem