Fully functional tilt only controls implemented
Included UI Texts at the bottom of the game screen to serve as in-game
debug text statements for ball, paddle positions and accelerometer
input. For now it looks kind of messing, since the numbers are being
rounded to 7 decimal places. I’ll just have to go back and round to
smaller values.