Added in menu icons
Included menu icons for the main menu and stuck the rotation script I
made for the title model onto the menu icons. Menu icons can now rotate
now! I just have to figure out how to make it so that it'll stop
automatically when it comes close to one of the icons