Added 2D sprite of hologram base
Added a sprite that will fade into view as the projector "powers up" for
added aesthetics. Also tried adding in functionality for when the player
taps on a menu option, the camera would rotate in the direction it needs
to go depending on the option. It doesn't work at the moment, because I
don't have colliders on the  menu options just yet. I'm not sure if I
have to add those the hard way though.